editing styling of tone rating slider
